added status and fixed issue 1
