+++ tasks_delete.go
@@ -24,7 +24,7 @@ func deleteTasks() {
 	}
 
 	// Create the AWS Service
-	svc := databasemigrationservice.New(s, &aws.Config{Region: aws.String("eu-west-1")})
+	svc := databasemigrationservice.New(s, &aws.Config{Region: &region})
 
 	// Read the defaults file
 	readTasks, err := ioutil.ReadFile(tasksFile)
@@ -34,31 +34,38 @@ func deleteTasks() {
 
 	// Create tasks and unmarshal the JSON
 	tasks := new([]ReplicationTask)
+	remainingTasks := new([]ReplicationTask) // Tasks that will be saved (if they couldn't be removed for example)
 	err = json.Unmarshal(readTasks, tasks)
 	if err != nil {
 		log.Fatal("Couldn't JSON unmarshal file "+tasksFile, err)
 	}
 
 	// Start all the tasks stored in tasks
-	for id, task := range *tasks {
+	for _, task := range *tasks {
 		params := &databasemigrationservice.DeleteReplicationTaskInput{
 			ReplicationTaskArn: aws.String(task.ReplicationTaskArn),
 		}
 
 		_, err := svc.DeleteReplicationTask(params)
 		if err != nil {
+			// If the task doesn't exists we shouldn't keep it in the tasks.json file - just continue
+			if strings.Contains(err.Error(), notFound) {
+				continue
+			}
+
+			// If the task errored and not because it doesn't exists, keep it in the tasks.json file
+			*remainingTasks = append(*remainingTasks, task)
+
+			// Go through the different statuses that might have made the action failed
 			switch {
-			case strings.Contains(err.Error(), "is running"):
-				fmt.Println("Please stop task", task.ReplicationTaskIdentifier, "before trying to delete it")
+			case strings.Contains(err.Error(), isRunning):
+				fmt.Println("Please stop task", task.ReplicationTaskIdentifier, stopBeforeDeleting)
 				continue
 			case strings.Contains(err.Error(), "is currently being stopped"):
-				fmt.Println("Please wait until task", task.ReplicationTaskIdentifier, "has stopped before trying to delete it")
+				fmt.Println("Please wait until task", task.ReplicationTaskIdentifier, waitForStop)
 				continue
 			case strings.Contains(err.Error(), "is already being deleted"):
-				fmt.Println("Task", task.ReplicationTaskIdentifier, "is already being deleted")
-				continue
-			case strings.Contains(err.Error(), "not found"):
-				removeTask(tasks, id)
+				fmt.Println("Task", task.ReplicationTaskIdentifier, beingDeleted)
 				continue
 			}
 
@@ -66,16 +73,13 @@ func deleteTasks() {
 			continue
 		}
 
-		// Remove task from tasks.json
-		removeTask(tasks, id)
-
 		counter++
 		fmt.Println("Task deleted: " + task.ReplicationTaskIdentifier)
 	}
 
 	// If we have no tasks left, delete the whole file
 	switch {
-	case len(*tasks) == 0:
+	case len(*remainingTasks) == 0:
 		err := os.Remove(tasksFile)
 		if err != nil {
 			fmt.Println("Couldn't remove tasks files", err)
@@ -83,18 +87,8 @@ func deleteTasks() {
 
 	default:
 		// Write remaining tasks to tasks-file
-		writeTaskFile(tasks)
+		writeTaskFile(remainingTasks)
 	}
 
 	fmt.Println("\nDONE! Deleted", counter, "tasks.")
-}
-
-func removeTask(tasks *[]ReplicationTask, id int) {
-	length := len(*tasks)
-	switch {
-	case length > id:
-		*tasks = append((*tasks)[:id], (*tasks)[id+1:]...)
-	case length == id:
-		tasks = new([]ReplicationTask)
-	}
 }
