+++ tasks_stop.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
+	"os"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -23,7 +24,7 @@ func stopTasks() {
 	}
 
 	// Create the AWS Service
-	svc := databasemigrationservice.New(s, &aws.Config{Region: aws.String("eu-west-1")})
+	svc := databasemigrationservice.New(s, &aws.Config{Region: &region})
 
 	// Read the defaults file
 	readTasks, err := ioutil.ReadFile(tasksFile)
@@ -33,37 +34,59 @@ func stopTasks() {
 
 	// Create tasks and unmarshal the JSON
 	tasks := new([]ReplicationTask)
+	remainingTasks := new([]ReplicationTask) // Tasks that will be saved (if they couldn't be removed for example)
 	err = json.Unmarshal(readTasks, tasks)
 	if err != nil {
 		log.Fatal("Couldn't JSON unmarshal file "+tasksFile, err)
 	}
 
 	// Start all the tasks stored in tasks
-	for id, task := range *tasks {
+	for _, task := range *tasks {
 		params := &databasemigrationservice.StopReplicationTaskInput{
 			ReplicationTaskArn: aws.String(task.ReplicationTaskArn),
 		}
 
 		_, err := svc.StopReplicationTask(params)
 		if err != nil {
+			// If the task doesn't exists we shouldn't keep it in the tasks.json file - just continue
+			if strings.Contains(err.Error(), doesntExists) {
+				continue
+			}
+
+			// If the task errored and not because it doesn't exists, keep it in the tasks.json file
+			*remainingTasks = append(*remainingTasks, task)
+
 			switch {
-			case strings.Contains(err.Error(), "is not currently running"):
+			case strings.Contains(err.Error(), currentlyNotRunning):
 				fmt.Println("Task", task.ReplicationTaskIdentifier, "is currently not running")
 				continue
-			case strings.Contains(err.Error(), "is already being stopped"):
+			case strings.Contains(err.Error(), allreadyBeingStopped):
 				fmt.Println("Task", task.ReplicationTaskIdentifier, "is already being stopped")
 				continue
-			case strings.Contains(err.Error(), "does not exist"):
-				fmt.Println("Task", task.ReplicationTaskIdentifier, "doesn't exists")
-				removeTask(tasks, id)
 			}
 
 			fmt.Println("Couldn't stop Replication Task", err)
 			continue
 		}
 
+		// If the task errored and not because it doesn't exists, keep it in the tasks.json file
+		*remainingTasks = append(*remainingTasks, task)
+
 		counter++
 		fmt.Println("Task stopped: " + task.ReplicationTaskIdentifier)
+	}
+
+	// If we have no tasks left, delete the whole file
+	switch {
+	case len(*remainingTasks) == 0:
+		err := os.Remove(tasksFile)
+		if err != nil {
+			fmt.Println("Couldn't remove tasks files", err)
+		}
+
+	default:
+		// Write remaining tasks to tasks-file
+		writeTaskFile(remainingTasks)
 	}
 
 	fmt.Println("\nDONE! Stopped", counter, "tasks.")
